Bugs fixed and main is divided for modules
